pet_name_generator/requirements.txt
langchain>=0.0.305,<0.1
openai>=0.28.1,<0.29
streamlit>=1.27.0,<1.28
python-dotenv>=1.0.0,<1.1
wikipedia>=1.4.0,<1.5